Moved container building into HTTP container builder. Eliminated Protean Container Builder.
