--- http/fab/Prefab5/Protean/Container/Builder/AwareTrait.php
@@ -1,45 +1,0 @@
-<?php
-declare(strict_types=1);
-
-namespace ReplaceThisWithTheNameOfYourVendor\ReplaceThisWithTheNameOfYourProduct\Prefab5\Protean\Container\Builder;
-
-use ReplaceThisWithTheNameOfYourVendor\ReplaceThisWithTheNameOfYourProduct\Prefab5\Protean\Container\BuilderInterface;
-
-trait AwareTrait
-{
-    protected $ReplaceThisWithTheNameOfYourVendorReplaceThisWithTheNameOfYourProductProteanContainerBuilder;
-
-    public function setProteanContainerBuilder(BuilderInterface $proteanContainerBuilder): self
-    {
-        if ($this->hasProteanContainerBuilder()) {
-            throw new \LogicException('ReplaceThisWithTheNameOfYourVendorReplaceThisWithTheNameOfYourProductProteanContainerBuilder is already set.');
-        }
-        $this->ReplaceThisWithTheNameOfYourVendorReplaceThisWithTheNameOfYourProductProteanContainerBuilder = $proteanContainerBuilder;
-
-        return $this;
-    }
-
-    protected function getProteanContainerBuilder(): BuilderInterface
-    {
-        if (!$this->hasProteanContainerBuilder()) {
-            throw new \LogicException('ReplaceThisWithTheNameOfYourVendorReplaceThisWithTheNameOfYourProductProteanContainerBuilder is not set.');
-        }
-
-        return $this->ReplaceThisWithTheNameOfYourVendorReplaceThisWithTheNameOfYourProductProteanContainerBuilder;
-    }
-
-    protected function hasProteanContainerBuilder(): bool
-    {
-        return isset($this->ReplaceThisWithTheNameOfYourVendorReplaceThisWithTheNameOfYourProductProteanContainerBuilder);
-    }
-
-    protected function unsetProteanContainerBuilder(): self
-    {
-        if (!$this->hasProteanContainerBuilder()) {
-            throw new \LogicException('ReplaceThisWithTheNameOfYourVendorReplaceThisWithTheNameOfYourProductProteanContainerBuilder is not set.');
-        }
-        unset($this->ReplaceThisWithTheNameOfYourVendorReplaceThisWithTheNameOfYourProductProteanContainerBuilder);
-
-        return $this;
-    }
-}
